requirements.txt
Flask==0.12
Flask-Assets==0.12
Flask-Cache==0.13.1
Flask-Cors==3.0.2
Flask-DebugToolbar==0.10.0
Flask-OpenID==1.2.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.14
Flask-SSLify==0.1.5
Jinja2==2.9.4
Fabric3==1.13.1.post1
celery==4.1.0
cssmin==0.2.0
jsmin==2.2.1
profanity==1.1
ts3==1.0.6
pyOpenSSL==16.2.0
python-dateutil==2.6.0
python3-openid==3.1.0
termcolor==1.1.0
